Moved the `layout` util to `LayoutManagement` in CARTs core utilities, in preparation for a unified layout handler.
