+++ CART/CARTLib/examples/MultiContrastSegmentation/MultiContrastSegmentationEvaluationTask.py
@@ -9,7 +9,7 @@ from slicer.i18n import tr as _
 
 from CARTLib.core.TaskBaseClass import TaskBaseClass, DataUnitFactory
 from CARTLib.utils.widgets import CARTSegmentationEditorWidget, showSuccessPrompt, showErrorPrompt
-from CARTLib.utils.layout import Orientation
+from CARTLib.core.LayoutManagement import Orientation
 from CARTLib.utils.config import ProfileConfig
 from CARTLib.utils.task import cart_task
 
@@ -139,8 +139,6 @@ class MultiContrastSegmentationEvaluationGUI:
 
     def _buildOutputModePrompt(self):
         """Build the output mode selection dialog with CSV logging option."""
-        from datetime import datetime
-        import csv
 
         prompt = qt.QDialog()
         prompt.setWindowTitle("Select Output Mode & Logging")
